09-06-2024 actualizacion con el crud completo y imagen

+++ app/Livewire/CategoriasComponent.php
@@ -23,9 +23,8 @@ class CategoriasComponent extends Component
     {
         $this->tituloPagina = 'Listado';
         $this->componenteNombre = 'Categorias';
-      
     }
-   
+
     // metodo para personalizar paginacion
     public function paginationView()
     {
@@ -49,67 +48,9 @@ class CategoriasComponent extends Component
         //return view('livewire.categorias', ['categorias' => $categorias]);
     }
 
-    //----Funcion para editar
-    //aqui se recibe el id enviado desde el click para buscar la informacion que se va a mostrar en el modal.. pero aun no actualiza
-    public function editar($id){
-        // esto trae todos los datos y funciona pero como buenas practicas podemos optimizar
-        //$grabar = Categoria::find($id);
-        $grabar = Categoria::find($id, ['id', 'nombre', 'imagen']);  //->metodo optimizado
-        $this->nombre = $grabar->nombre;
-        $this->seleccionar_id = $grabar->id;
-        $this->imagen = null;
-        // este es un evento emitido desde el backend y este debe escucharse desde el frontend con javascrip en este caso
-       // $this->emit('show-modal', 'show modal!'); --> metodo anterior
-        $this->dispatch('show-modal','show modal!'); //-->> metodo actualizado,, para mostrar el modal
-    }
-
-    // ------- Metodo actualizar categoria-------------
-        public function update(){
-            $rules = [
-                'nombre' => "required|min:3|unique:categorias,nombre,{$this->seleccionar_id}"
-            ];
-            $mensaje =[
-                'nombre.required' => 'Debe ingrear el nombre',
-                'nombre.min' => 'El nombre minimo debe tener 3 caracteres',
-                'nombre.unique'=>'El nnombre de la categoria ya existe'
-            ];
-            $this->validate($rules , $mensaje);
-
-            //busca la categoria segun el codigo
-            $categoria = Categoria::find($this->seleccionar_id);
-            $categoria->update([
-                'nombre' => $this->nombre
-            ]);
-
-            if($this->imagen){
-                // el metodo uniquid para dar un valor unico..
-                $archivocliente = uniqid() . '_.' . $this->imagen->extension();
-                //ruta donde se va depositar la imagen
-                $this->imagen->storeAs('public/categorias',$archivocliente);
-
-                // como se va actuyalizar la imagen es necesario borrar la exsitente
-                $imagennombre = $categoria->imagen; // recupera la imagen que esta guardada
-                $categoria->imagen = $archivocliente; // aca se le asigna la nueva imagen
-                $categoria->save();
-
-                if($imagennombre != null){
-                    if(file_exists('storage/categorias/' . $imagennombre)){
-                        unlink('storage/categorias/' . $imagennombre);
-                    }
-                }
-
-                //--limpiar
-                $this->limpiar();
-                //evento para mostrar el mensaje y cerrar el modal
-                $this->dispatch('categoria-actualizada','Categoria Actualizada');
-
-            };
-
-        }
-
-
-    //-----metodo para registrar las categorias
-    public function store(){
+    //-----metodo para registrar las categorias------------------------------------------------------
+    public function store()
+    {
         $rules = [
             'nombre' => 'required|unique:categorias|min:3'
         ];
@@ -118,7 +59,7 @@ class CategoriasComponent extends Component
             'nombre.unique' => 'La información no puede ser duplicada',
             'nombre.min' => 'Catacteres insuficientes',
         ];
-    
+
 
         $this->validate($rules, $mensajes);
 
@@ -140,6 +81,110 @@ class CategoriasComponent extends Component
         //evento para mostrar el mensaje y cerrar el modal si es caso
         $this->dispatch('categoria-agregada', 'Categoriar Registrada');
     }
+
+
+    //----Funcion para editar------------------------------------------
+    //aqui se recibe el id enviado desde el click para buscar la informacion que se va a mostrar en el modal.. pero aun no actualiza
+    public function editar($id)
+    {
+        // esto trae todos los datos y funciona pero como buenas practicas podemos optimizar
+        //$grabar = Categoria::find($id);
+        $grabar = Categoria::find($id, ['id', 'nombre', 'imagen']);  //->metodo optimizado
+        $this->nombre = $grabar->nombre;
+        $this->seleccionar_id = $grabar->id;
+        $this->imagen = null;
+        // este es un evento emitido desde el backend y este debe escucharse desde el frontend con javascrip en este caso
+        // $this->emit('show-modal', 'show modal!'); --> metodo anterior
+        $this->dispatch('show-modal', 'show modal!'); //-->> metodo actualizado,, para mostrar el modal
+    }
+
+    // ------- Metodo actualizar categoria-------------
+    // public function update(){
+    //     $rules = [
+    //         'nombre' => "required|min:3|unique:categorias,nombre,{$this->seleccionar_id}"
+    //     ];
+    //     $mensaje =[
+    //         'nombre.required' => 'Debe ingrear el nombre',
+    //         'nombre.min' => 'El nombre minimo debe tener 3 caracteres',
+    //         'nombre.unique'=>'El nnombre de la categoria ya existe'
+    //     ];
+    //     $this->validate($rules , $mensaje);
+
+    //     //busca la categoria segun el codigo
+    //     $categoria = Categoria::find($this->seleccionar_id);
+    //     $categoria->update([
+    //         'nombre' => $this->nombre
+    //     ]);
+
+    //     if($this->imagen){
+    //         // el metodo uniquid para dar un valor unico..
+    //         $archivocliente = uniqid() . '_.' . $this->imagen->extension();
+    //         //ruta donde se va depositar la imagen
+    //         $this->imagen->storeAs('public/categorias',$archivocliente);
+
+    //         // como se va actuyalizar la imagen es necesario borrar la exsitente
+    //         $imagennombre = $categoria->imagen; // recupera la imagen que esta guardada
+    //         $categoria->imagen = $archivocliente; // aca se le asigna la nueva imagen
+    //         $categoria->save();
+
+    //         if($imagennombre != null){
+    //             if(file_exists('storage/categorias/' . $imagennombre)){
+    //                 unlink('storage/categorias/' . $imagennombre);
+    //             }
+    //         }
+
+    //         //--limpiar
+    //         $this->limpiar();
+    //         //evento para mostrar el mensaje y cerrar el modal
+    //         $this->dispatch('categoria-actualizada','Categoria Actualizada');
+
+    //     };
+
+    // }
+
+    public function update()
+    {
+        $rules = [
+            'nombre' => "required|min:3|unique:categorias,nombre,{$this->seleccionar_id}"
+        ];
+        $mensaje = [
+            'nombre.required' => 'Debe ingresar el nombre',
+            'nombre.min' => 'El nombre mínimo debe tener 3 caracteres',
+            'nombre.unique' => 'El nombre de la categoría ya existe'
+        ];
+        $this->validate($rules, $mensaje);
+
+        // Busca la categoría según el código
+        $categoria = Categoria::find($this->seleccionar_id);
+        $categoria->update([
+            'nombre' => $this->nombre
+        ]);
+
+        if ($this->imagen) {
+            // El método uniqid para dar un valor único
+            $archivocliente = uniqid() . '_.' . $this->imagen->extension();
+            // Ruta donde se va a depositar la imagen
+            $this->imagen->storeAs('public/categorias', $archivocliente);
+
+            // Como se va a actualizar la imagen, es necesario borrar la existente
+            $imagennombre = $categoria->getImagenPath(); // Recupera el nombre del archivo de la imagen que está guardada
+            $categoria->imagen = $archivocliente; // Asigna la nueva imagen
+            $categoria->save();
+
+            if ($imagennombre != null) {
+                if (file_exists(public_path('storage/categorias/' . $imagennombre))) {
+                    unlink(public_path('storage/categorias/' . $imagennombre));
+                }
+            }
+
+            // Limpiar
+            $this->limpiar();
+            // Evento para mostrar el mensaje y cerrar el modal
+            $this->dispatch('categoria-actualizada', 'Categoria Actualizada');
+        }
+    }
+
+
     public function limpiar()
     {
         $this->nombre = "";
@@ -149,36 +194,63 @@ class CategoriasComponent extends Component
 
     // en este caso el evento se esta desarrollando desde el fronend y debe ser escuchado en el backend
     // funcion para escuchar el evento
-        public $listeners = [
-            'eliminarFila' => 'destroy' // el id lo pasa automaticamente
-        ];
+
+    protected $listeners = ['eliminarFila'];
+
+    //esta funcion puede ser un poco mas optimisada ejemplo:
+    // public function eliminarFila(Categoria $categoria) {
+    //  y se elimina este codigo $categoria = Categoria::find($id);  }       
+
+    // public function eliminarFila($id)
+    // {
+    //     $categoria = Categoria::find($id);
+    //    // dd($categoria);
+
+    //     if ($categoria) {
+    //         $nombreimagen = $categoria->imagen;
+    //         $categoria->delete();  // hasta aca ya se elimina el registro de la base de datos con el metodo delete de elocuent...
+    //         //pero falta eliminar la imagne del del rpoyecyo o disco para no tener archivos basuras.
 
 
-//         public function mount()
-// {
-//   $this->listen('eliminarFila', 'destroy'); // Escucha el evento y enlázalo al método
-// }
+    //         //1) se valida si esta en la baase de datos 
+    //         if ($nombreimagen != null) {
+    //             // se utilizar el metodo unlink para eliminar la imgen que esta con la ruta dada
+    //             unlink('storage/categorias/' . $nombreimagen);                   
+    //         }
 
+    //         $this->limpiar();
+    //         //$this->dispatchBrowserEvent('categoria-eliminada', ['message' => 'Categoria eliminada']);
+    //         $this->dispatch('categoria-eliminada','Categoria eliminada con exito');
 
+    //     } else {
+    //         // Manejar el caso cuando la categoría no se encuentra
+    //         //$this->dispatchBrowserEvent('categoria-no-encontrada', ['message' => 'Categoria no encontrada']);
+    //         $this->dispatch('categoria-no-encontrada','Categoria no encontrada');
+    //     }
+    // }
 
+    public function eliminarFila($id)
+    {
+        $categoria = Categoria::find($id);
+        // dd($categoria);
 
-        //---Funcion para eliminar categorias
-        public function destroy($id){
-            //buscar el id por medio de la peticion
-            $categoria = Categoria::find($id);
-            dd($categoria);
-            //variable para almacenar la imagen que hay actualmente
-            $nombreimagen = $categoria->imagen;
-            $categoria->delete(); // hsta aca se elimina la imagen pero la imagen queda en nuestros archivos
-        
-            // eliminar la imagen de nuestro archivo
-            if($nombreimagen != null ){
-                //unlick metodo para eliminar la imagen en la ruta dada
-                unlink('storage/categorias/' . $nombreimagen);
+        if ($categoria) {
+            // Usar el método getImagenPath() para obtener solo el nombre del archivo
+            $nombreimagen = $categoria->getImagenPath();
+            $categoria->delete();  // Eliminar el registro de la base de datos
+
+            // Validar si el nombre de la imagen no es nulo y eliminar la imagen del sistema de archivos
+            if ($nombreimagen != null && file_exists(public_path('storage/categorias/' . $nombreimagen))) {
+                unlink(public_path('storage/categorias/' . $nombreimagen));
             }
 
             $this->limpiar();
-            //evento para mostrar el mensaje y cerrar el modal
-            $this->dispatch('categoria-eiminada','Categoria eliminada');
+            // $this->dispatchBrowserEvent('categoria-eliminada', ['message' => 'Categoria eliminada']);
+            $this->dispatch('categoria-eliminada', 'Categoria eliminada con exito');
+        } else {
+            // Manejar el caso cuando la categoría no se encuentra
+            // $this->dispatchBrowserEvent('categoria-no-encontrada', ['message' => 'Categoria no encontrada']);
+            $this->dispatch('categoria-no-encontrada', 'Categoria no encontrada');
         }
-}   
+    }
+}
